fix: move installActivityLog into helpers (ROK-1116)
CI lint failed because demo-data.service.ts crossed the 300-line
ESLint cap (which is `error`, not `warn`, contrary to CLAUDE.md).
Pre-rebase the file was 296 LOC; the new private method tipped it
to 316.

Move the orchestration logic next to the row builders in
demo-data-install-activity.helpers.ts where it belongs anyway,
matching the demo-data-install-{core,signups,secondary,taste}
helper convention.

Behavior is unchanged — same re-fetch of post-reassignment creators,
same single bulk insert. All 12 demo-data integration tests pass.
