+++ api/src/admin/demo-data-install-activity.helpers.ts
@@ -5,7 +5,16 @@
  * normally call ActivityLogService.log(...). These pure builders produce the
  * activity_log inserts so demo events render proper timelines.
  */
+import { inArray } from 'drizzle-orm';
+import type { PostgresJsDatabase } from 'drizzle-orm/postgres-js';
 import * as schema from '../drizzle/schema';
+
+type Db = PostgresJsDatabase<typeof schema>;
+type BatchInsert = (
+  table: Parameters<Db['insert']>[0],
+  rows: Record<string, unknown>[],
+  onConflict?: 'doNothing',
+) => Promise<unknown>;
 
 type EventRow = Pick<
   typeof schema.events.$inferSelect,
@@ -44,4 +53,28 @@ export function buildSignupAddedActivityRows(
         metadata: { role: roles?.[0] ?? null },
       };
     });
+}
+
+/** Re-fetches creator_id post-reassignment, then bulk-inserts activity rows. */
+export async function installActivityLog(
+  db: Db,
+  batchInsert: BatchInsert,
+  createdEvents: (typeof schema.events.$inferSelect)[],
+  createdSignups: (typeof schema.eventSignups.$inferSelect)[],
+): Promise<void> {
+  if (createdEvents.length === 0) return;
+  const eventIds = createdEvents.map((e) => e.id);
+  const finalEvents = await db
+    .select({
+      id: schema.events.id,
+      creatorId: schema.events.creatorId,
+      title: schema.events.title,
+    })
+    .from(schema.events)
+    .where(inArray(schema.events.id, eventIds));
+  const rows = [
+    ...buildEventCreatedActivityRows(finalEvents),
+    ...buildSignupAddedActivityRows(createdSignups),
+  ];
+  if (rows.length > 0) await batchInsert(schema.activityLog, rows);
 }
